requirements.txt
beautifulsoup4==4.3.2
cmsplugin-iframe==0.1.1
dj-database-url==0.3.0
Django==1.7.7
django-appconf==1.0.1
django-classy-tags==0.6.1
django-cms==3.0.12
django-haystack==2.3.1
django-mptt==0.6.1
django-reversion==1.8.5
django-sekizai==0.8.1
django-sortedm2m==0.9.4
django-spurl==0.6
django-standard-form==1.1.1
django-taggit==0.12.3
djangocms-admin-style==0.2.5
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.2
djangocms-link==1.5
djangocms-picture==0.1
djangocms-snippet==1.4
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
html5lib==0.999
lxml==3.4.2
Pillow==2.8.1
pysolr==3.3.0
pytz==2014.10
requests==2.6.0
six==1.9.0
South==1.0.2
URLObject==2.4.0
YURL==0.13
